+++ shared/src/commonMain/kotlin/io/madrona/njord/model/Chart.kt
@@ -9,7 +9,6 @@ import kotlinx.serialization.json.JsonElement
 
 @Serializable
 class ChartInfo(
-    val id: Long,
     val name: String,
     val scale: Int,
     val zoom: Int,
@@ -33,7 +32,6 @@ class ChartFeature(
 
 @Serializable
 data class Chart(
-    val id: Long,
     val name: String,
     val scale: Int,
     @SerialName("file_name") val fileName: String,